Externalize audio helpers
