+++ index.html
@@ -123,6 +123,6 @@
     </div>
   </div>
 
-  <script src="tetris.js"></script>
+  <script type="module" src="src/main.js"></script>
 </body>
 </html>
